Force generate zones with pyramids
This is so that zones that were never scanned do not report a white pyramid.

Unfortunately at the time of writing there is no way to trigger Zone.delete_surface from an interface or automatically via the gui, so the player will unfortunately be forced to do some manual labour.